+++ package.json
@@ -18,7 +18,7 @@
   "dependencies": {
     "@mdx-js/react": "^2.1.5",
     "@primer/octicons-react": "^17.9.0",
-    "gatsby": "^5.3.2",
+    "gatsby": "^5.3.3",
     "gatsby-plugin-mdx": "^5.0.0",
     "gatsby-source-filesystem": "^5.0.0",
     "react": "^18.2.0",
