+++ src/components/header.tsx
@@ -3,6 +3,7 @@ import { Link } from "gatsby";
 import Logo from "../images/logo.png";
 import "./header.css";
 
+import { MarkGithubIcon, MailIcon } from "@primer/octicons-react";
 interface Props {
     hide?: boolean,
     text?: string,
@@ -20,6 +21,19 @@ export default function Header({ hide, text }: Props) {
         <ul className="header">
             <li><Link to="/"><img src={Logo} alt="logo" /></Link></li>
             <li><h1>{text}</h1></li>
+            <div className="socials">
+                <ul>
+                    <li><a href="mailto:ourvoicetechnologies@gmail.com"><MailIcon verticalAlign="middle" /></a></li>
+                    <li><a href="https://github.com/ourvoicetechnologies"><MarkGithubIcon verticalAlign="middle" /></a></li>
+                </ul>
+            </div>
+            <div className="nav">
+                <ul>
+                    <li><Link to="/about">About</Link></li>
+                    <li><Link to="/ourvoiceforstudents">OurVoice For Students</Link></li>
+                    <li><Link to="/contact">Contact Info</Link></li>
+                </ul>
+            </div>
         </ul>
     );
 }
