+++ src/components/layout.css
@@ -8,6 +8,21 @@ body {
     /* font-family: Arial, Helvetica, sans-serif; */
 }
 
+main {
+    animation: fadeInAnimation ease 2s;
+    animation-iteration-count: 1;
+    animation-fill-mode: forwards;
+}
+  
+@keyframes fadeInAnimation {
+    0% {
+        opacity: 0;
+    }
+    100% {
+        opacity: 1;
+     }
+} 
+
 .center {
     display: inline-block;
     text-align: center;
